refactor: Create the domain for checkout and discount rules
